% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{BIC.mixLCA}
\alias{BIC.mixLCA}
\title{BIC Method for mixLCA}
\usage{
\method{BIC}{mixLCA}(object, ...)
}
\arguments{
\item{object}{A \code{mixLCA} object.}

\item{...}{Unused.}
}
\value{
Numeric BIC value.
}
\description{
BIC Method for mixLCA
}
