% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_distal.R
\name{classification_error_matrix}
\alias{classification_error_matrix}
\title{Compute BCH Classification Error Matrix}
\usage{
classification_error_matrix(posteriors)
}
\arguments{
\item{posteriors}{N x K posterior probability matrix.}
}
\value{
K x K classification error matrix.
}
\description{
Constructs the K x K matrix E where E[w, k] = P(W = w | C = k),
i.e. the probability that an observation truly belonging to class k
is modally assigned to class w.
}
\keyword{internal}
