% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{coef.mixLCA}
\alias{coef.mixLCA}
\title{Coef Method for mixLCA}
\usage{
\method{coef}{mixLCA}(object, ...)
}
\arguments{
\item{object}{A \code{mixLCA} object.}

\item{...}{Unused.}
}
\value{
Named list.
}
\description{
Returns a list of all estimated parameters.
}
