% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concomitant.R
\name{compute_priors}
\alias{compute_priors}
\title{Compute Prior Probabilities from the Concomitant Model}
\usage{
compute_priors(X, coefs)
}
\arguments{
\item{X}{Numeric matrix (N x P), includes intercept.}

\item{coefs}{P x (K-1) coefficient matrix.}
}
\value{
Numeric matrix (N x K) of prior class probabilities.
}
\description{
Compute Prior Probabilities from the Concomitant Model
}
\keyword{internal}
