% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_concomitant.R
\name{concomitant_grad}
\alias{concomitant_grad}
\title{Analytical Gradient for Concomitant Model}
\usage{
concomitant_grad(par, X, posteriors)
}
\arguments{
\item{par}{Numeric vector of length P*(K-1).}

\item{X}{Numeric matrix (N x P).}

\item{posteriors}{Numeric matrix (N x K).}
}
\value{
Numeric vector (gradient of the negative log-likelihood).
}
\description{
Analytical Gradient for Concomitant Model
}
\keyword{internal}
