% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_standard_errors.R
\name{concomitant_se}
\alias{concomitant_se}
\title{Standard Errors for Concomitant Coefficients}
\usage{
concomitant_se(model, data)
}
\arguments{
\item{model}{A \code{mixLCA} object.}

\item{data}{Data frame used for estimation.}
}
\value{
A P x (K-1) matrix of standard errors, or NULL if no
  concomitant predictors were specified.
}
\description{
Constructs the empirical (observed) information matrix using the
outer product of per-observation score vectors, then inverts to
obtain asymptotic variances for the multinomial logistic coefficients.
}
\examples{
\donttest{
data(health_screening)
fit <- fit_lca(health_screening,
               continuous  = c("marker_1","marker_2","marker_3","marker_4"),
               concomitant = ~ age,
               n_classes   = 2,
               control     = lca_control(n_starts = 3),
               verbose     = FALSE)
se <- concomitant_se(fit, health_screening)
round(cbind(Estimate = fit$concomitant_coefs[, 1], SE = se[, 1]), 4)
}
}
