% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{count_params}
\alias{count_params}
\title{Count Free Parameters in a mixLCA Object}
\usage{
count_params(model)
}
\arguments{
\item{model}{A \code{mixLCA} object.}
}
\value{
Integer.
}
\description{
Tallies the number of estimable parameters.
}
\keyword{internal}
