% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_spectral_dependence.R
\name{eval_spectral_density}
\alias{eval_spectral_density}
\title{Evaluate Spectral Composite Log-Density}
\usage{
eval_spectral_density(spec_data, pi_c, A_star)
}
\description{
Evaluate Spectral Composite Log-Density
}
\keyword{internal}
