% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_spectral_dependence.R
\name{flatten_cat_params}
\alias{flatten_cat_params}
\title{Flatten Categorical Parameters List}
\usage{
flatten_cat_params(cat_params_k, item_indices, C)
}
\description{
Flatten Categorical Parameters List
}
\keyword{internal}
