% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{get_posteriors}
\alias{get_posteriors}
\alias{get_posteriors.mixLCA}
\title{Extract Posterior Class Probabilities}
\usage{
get_posteriors(x, ...)

\method{get_posteriors}{mixLCA}(x, ...)
}
\arguments{
\item{x}{A fitted \code{mixLCA} object.}

\item{...}{Unused.}
}
\value{
Numeric matrix.
}
\description{
Accessor returning the N x K posterior probability matrix. Use this
rather than \code{model$posteriors} so downstream code is insulated
against future restructuring of internal fields.
}
\examples{
\donttest{
data(voter_perceptions)
fit <- fit_lca(voter_perceptions,
               categorical = names(voter_perceptions),
               n_classes   = 2,
               control     = lca_control(n_starts = 2),
               verbose     = FALSE)
post <- get_posteriors(fit)
dim(post)
head(round(post, 3))
colSums(post) / nrow(post)   # estimated class proportions
}
}
