% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Numerically Stable Log-Sum-Exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Scalar.
}
\description{
Computes log(sum(exp(x))) without overflow or underflow.
}
\keyword{internal}
