% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{logLik.mixLCA}
\alias{logLik.mixLCA}
\title{logLik Method for mixLCA}
\usage{
\method{logLik}{mixLCA}(object, ...)
}
\arguments{
\item{object}{A \code{mixLCA} object.}

\item{...}{Unused.}
}
\value{
Object of class \code{logLik}.
}
\description{
logLik Method for mixLCA
}
\examples{
\donttest{
data(voter_perceptions)
fit <- fit_lca(voter_perceptions,
               categorical = names(voter_perceptions),
               n_classes   = 2,
               control     = lca_control(n_starts = 2),
               verbose     = FALSE)
logLik(fit)
}
}
