% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixLCA-package.R
\docType{package}
\name{mixLCA-package}
\alias{mixLCA}
\alias{mixLCA-package}
\title{mixLCA: Partitioned Latent Class Analysis}
\description{
Finite mixture modelling with partitioned architecture for antecedent
predictors, contemporaneous manifest indicators, and distal outcomes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://asanaei.github.io/mixLCA/}
  \item \url{https://github.com/asanaei/mixLCA}
  \item Report bugs at \url{https://github.com/asanaei/mixLCA/issues}
}

}
\author{
\strong{Maintainer}: Ali Sanaei \email{sanaei@uchicago.edu}

}
