% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{nobs.mixLCA}
\alias{nobs.mixLCA}
\title{nobs Method for mixLCA}
\usage{
\method{nobs}{mixLCA}(object, ...)
}
\arguments{
\item{object}{A \code{mixLCA} object.}

\item{...}{Unused.}
}
\value{
Integer.
}
\description{
nobs Method for mixLCA
}
