% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_spectral_dependence.R
\name{prep_spectral_data}
\alias{prep_spectral_data}
\title{Prepare Categorical Data for Spectral Local Dependence}
\usage{
prep_spectral_data(df)
}
\description{
Prepare Categorical Data for Spectral Local Dependence
}
\keyword{internal}
