% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{print.mixLCA}
\alias{print.mixLCA}
\title{Print a mixLCA Object}
\usage{
\method{print}{mixLCA}(x, ...)
}
\arguments{
\item{x}{A \code{mixLCA} object.}

\item{...}{Unused.}
}
\description{
Print a mixLCA Object
}
\examples{
\donttest{
data(voter_perceptions)
fit <- fit_lca(voter_perceptions,
               categorical = names(voter_perceptions),
               n_classes   = 2,
               control     = lca_control(n_starts = 2, seed = 110),
               verbose     = FALSE)
print(fit)
}
}
