% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{softmax}
\alias{softmax}
\title{Softmax Transformation}
\usage{
softmax(X)
}
\arguments{
\item{X}{Numeric matrix (N x K).}
}
\value{
Matrix of probabilities (N x K).
}
\description{
Row-wise softmax of a matrix. Each row sums to 1.
}
\keyword{internal}
