% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_spectral_dependence.R
\name{spectral_loadings}
\alias{spectral_loadings}
\title{Spectral Loadings as a Tidy Data Frame}
\usage{
spectral_loadings(model)
}
\arguments{
\item{model}{A \code{mixLCA} object.}
}
\value{
Data frame.
}
\description{
Spectral Loadings as a Tidy Data Frame
}
\examples{
\donttest{
data(voter_perceptions)
fit <- fit_lca(voter_perceptions,
               categorical   = names(voter_perceptions),
               n_classes     = 2,
               spectral_rank = c(1L, 1L),
               control       = lca_control(n_starts = 2),
               verbose       = FALSE)
ld <- spectral_loadings(fit)
head(ld, 12)
}
}
