% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_spectral_dependence.R
\name{spectral_projector}
\alias{spectral_projector}
\title{Spectral Projector Calculation}
\usage{
spectral_projector(Sigma_c, d, M)
}
\description{
Spectral Projector Calculation
}
\keyword{internal}
