% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{AIC.mixLCA}
\alias{AIC.mixLCA}
\title{AIC Method for mixLCA}
\usage{
\method{AIC}{mixLCA}(object, ..., k = 2)
}
\arguments{
\item{object}{A \code{mixLCA} object.}

\item{...}{Unused.}

\item{k}{Numeric penalty per parameter (default 2).}
}
\value{
Numeric AIC value.
}
\description{
AIC Method for mixLCA
}
