% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_s3_methods.R
\name{coef.mixLCA}
\alias{coef.mixLCA}
\title{Coef Method for mixLCA}
\usage{
\method{coef}{mixLCA}(object, ...)
}
\arguments{
\item{object}{A \code{mixLCA} object.}

\item{...}{Unused.}
}
\value{
Named list.
}
\description{
Returns a list of all estimated parameters.
}
\examples{
\donttest{
data(health_screening)
fit <- fit_lca(health_screening,
               continuous  = c("marker_1","marker_2","marker_3","marker_4"),
               concomitant = ~ age,
               n_classes   = 2,
               control     = lca_control(n_starts = 2),
               verbose     = FALSE)
cc <- coef(fit)
names(cc)
cc$concomitant
cc$means
}
}
