% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_spectral_dependence.R
\name{spectral_loadings}
\alias{spectral_loadings}
\title{Spectral Loadings as a Tidy Data Frame}
\usage{
spectral_loadings(model)
}
\arguments{
\item{model}{A \code{mixLCA} object.}
}
\value{
Data frame.
}
\description{
Spectral Loadings as a Tidy Data Frame
}
